+++ Source/PropertyTools.Wpf/DataGrid/CellDefinitions/CheckCellDefinition.cs
@@ -1,0 +1,19 @@
+// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="CheckCellDefinition.cs" company="PropertyTools">
+//   Copyright (c) 2014 PropertyTools contributors
+// </copyright>
+// <summary>
+//   Defines a cell that contains a boolean property.
+// </summary>
+// --------------------------------------------------------------------------------------------------------------------
+
+namespace PropertyTools.Wpf
+{
+    /// <summary>
+    /// Defines a cell that contains a boolean property.
+    /// </summary>
+    /// <seealso cref="PropertyTools.Wpf.CellDefinition" />
+    public class CheckCellDefinition : CellDefinition
+    {
+    }
+}
