+++ Source/PropertyTools.Wpf/DataGrid/CellDefinitions/ColorCellDefinition.cs
@@ -1,0 +1,18 @@
+// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="ColorCellDefinition.cs" company="PropertyTools">
+//   Copyright (c) 2014 PropertyTools contributors
+// </copyright>
+// <summary>
+//   Defines a cell that contains a color property.
+// </summary>
+// --------------------------------------------------------------------------------------------------------------------
+
+namespace PropertyTools.Wpf
+{
+    /// <summary>
+    /// Defines a cell that contains a color property.
+    /// </summary>
+    public class ColorCellDefinition : CellDefinition
+    {
+    }
+}
